refine parseMetadata
