+++ pkg/dumpling/utils.go
@@ -1,0 +1,168 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package dumpling
+
+import (
+	"bufio"
+	"fmt"
+	"io"
+	"os"
+	"strconv"
+	"strings"
+
+	"github.com/siddontang/go-mysql/mysql"
+
+	"github.com/pingcap/dm/pkg/binlog"
+	"github.com/pingcap/dm/pkg/gtid"
+	"github.com/pingcap/dm/pkg/terror"
+)
+
+// ParseMetaData parses mydumper's output meta file and returns binlog location.
+// since v2.0.0, dumpling maybe configured to output master status after connection pool is established,
+// we return this location as well.
+func ParseMetaData(filename, flavor string) (*binlog.Location, *binlog.Location, error) {
+	fd, err := os.Open(filename)
+	if err != nil {
+		return nil, nil, terror.ErrParseMydumperMeta.Generate(err)
+	}
+	defer fd.Close()
+
+	pos := mysql.Position{}
+	gtidStr := ""
+	pos2 := mysql.Position{}
+	gtidStr2 := ""
+
+	br := bufio.NewReader(fd)
+
+	parsePosAndGTID := func(pos *mysql.Position, gtid *string) error {
+		for {
+			line, err := br.ReadString('\n')
+			if err != nil {
+				return err
+			}
+			parts := strings.SplitN(line, ":", 2)
+			if len(parts) != 2 {
+				continue
+			}
+			key := strings.TrimSpace(parts[0])
+			value := strings.TrimSpace(parts[1])
+			switch key {
+			case "Log":
+				pos.Name = value
+			case "Pos":
+				pos64, err := strconv.ParseUint(value, 10, 32)
+				if err != nil {
+					return err
+				}
+				pos.Pos = uint32(pos64)
+			case "GTID":
+				// multiple GTID sets may cross multiple lines, continue to read them.
+				following, err := readFollowingGTIDs(br, flavor)
+				if err != nil {
+					return err
+				}
+				*gtid = value + following
+				return nil
+			}
+		}
+	}
+
+	for {
+		line, err := br.ReadString('\n')
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			return nil, nil, terror.ErrParseMydumperMeta.Generate(err)
+		}
+		line = strings.TrimSpace(line)
+		if len(line) == 0 {
+			continue
+		}
+
+		if line == "SHOW MASTER STATUS:" {
+			if err := parsePosAndGTID(&pos, &gtidStr); err != nil {
+				return nil, nil, terror.ErrParseMydumperMeta.Generate(err)
+			}
+			// ref: https://github.com/maxbube/mydumper/blob/master/mydumper.c#L434
+		} else if line == "SHOW SLAVE STATUS:" {
+			for {
+				line, err := br.ReadString('\n')
+				if err != nil {
+					return nil, nil, terror.ErrParseMydumperMeta.Generate(err)
+				}
+				line = strings.TrimSpace(line)
+				if len(line) == 0 {
+					break
+				}
+			}
+		} else if line == "SHOW MASTER STATUS: /* AFTER CONNECTION POOL ESTABLISHED */" {
+			if err := parsePosAndGTID(&pos2, &gtidStr2); err != nil {
+				return nil, nil, terror.ErrParseMydumperMeta.Generate(err)
+			}
+		}
+	}
+
+	if len(pos.Name) == 0 || pos.Pos == uint32(0) || len(pos2.Name) == 0 || pos2.Pos == uint32(0) {
+		return nil, nil, terror.ErrParseMydumperMeta.Generate(fmt.Sprintf("file %s invalid format", filename))
+	}
+
+	gset, err := gtid.ParserGTID(flavor, gtidStr)
+	if err != nil {
+		return nil, nil, terror.ErrParseMydumperMeta.Generate(fmt.Sprintf("file %s invalid format", filename))
+	}
+	loc := &binlog.Location{
+		Position: pos,
+		GTIDSet:  gset,
+	}
+	gset2, err := gtid.ParserGTID(flavor, gtidStr2)
+	if err != nil {
+		return nil, nil, terror.ErrParseMydumperMeta.Generate(fmt.Sprintf("file %s invalid format", filename))
+	}
+	loc2 := &binlog.Location{
+		Position: pos2,
+		GTIDSet:  gset2,
+	}
+
+	return loc, loc2, nil
+}
+
+func readFollowingGTIDs(br *bufio.Reader, flavor string) (string, error) {
+	var following strings.Builder
+	for {
+		line, err := br.ReadString('\n')
+		if err == io.EOF {
+			return following.String(), nil // return the previous, not including the last line.
+		} else if err != nil {
+			return "", err
+		}
+
+		line = strings.TrimSpace(line)
+		if len(line) == 0 {
+			return following.String(), nil // end with empty line.
+		}
+
+		end := len(line)
+		if strings.HasSuffix(line, ",") {
+			end = len(line) - 1
+		}
+
+		// try parse to verify it
+		_, err = gtid.ParserGTID(flavor, line[:end])
+		if err != nil {
+			return following.String(), nil // return the previous, not including this non-GTID line.
+		}
+
+		following.WriteString(line)
+	}
+}
